Added ability to use generic template based note-types.
Initially, there could only be 'notes', 'journals' and 'specs'.  With these changes, Tiro will search the template directory for any template that matches the note type specified on the command line.  So, issuing the command 'tiro foo Some text' will search the template directory for a template called 'foo_template' and use that as the base for the note.  Any note type that doesn't match a template will still be created, but without any template base.

Also updated the readme to reflect current usage.